// .PRD/erp-refactor/v3/types/purchase-order-item.ts
import { z } from 'zod';

export const PurchaseOrderItemSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  fg_item: z.string().optional(),
  fg_item_qty: z.number().optional().default(1),
  item_code: z.string(),
  supplier_part_no: z.string().optional(),
  item_name: z.string(),
  brand: z.string().optional(),
  product_bundle: z.string().optional(),
  schedule_date: z.string(),
  expected_delivery_date: z.string().optional(),
  item_group: z.string().optional(),
  description: z.string().optional(),
  image: z.string().optional(),
  image_view: z.string().optional(),
  qty: z.number(),
  stock_uom: z.string(),
  subcontracted_qty: z.number().optional().default(0),
  uom: z.string(),
  conversion_factor: z.number(),
  stock_qty: z.number().optional(),
  price_list_rate: z.number().optional(),
  last_purchase_rate: z.number().optional(),
  base_price_list_rate: z.number().optional(),
  margin_type: z.enum(['Percentage', 'Amount']).optional(),
  margin_rate_or_amount: z.number().optional(),
  rate_with_margin: z.number().optional(),
  discount_percentage: z.number().optional(),
  discount_amount: z.number().optional(),
  distributed_discount_amount: z.number().optional(),
  base_rate_with_margin: z.number().optional(),
  rate: z.number().optional(),
  amount: z.number().optional(),
  item_tax_template: z.string().optional(),
  base_rate: z.number(),
  base_amount: z.number(),
  pricing_rules: z.string().optional(),
  stock_uom_rate: z.number().optional(),
  is_free_item: z.boolean().optional().default(false),
  net_rate: z.number().optional(),
  net_amount: z.number().optional(),
  base_net_rate: z.number().optional(),
  base_net_amount: z.number().optional(),
  from_warehouse: z.string().optional(),
  warehouse: z.string().optional(),
  actual_qty: z.number().optional(),
  company_total_stock: z.number().optional(),
  material_request: z.string().optional(),
  material_request_item: z.string().optional(),
  sales_order: z.string().optional(),
  sales_order_item: z.string().optional(),
  sales_order_packed_item: z.string().optional(),
  supplier_quotation: z.string().optional(),
  supplier_quotation_item: z.string().optional(),
  delivered_by_supplier: z.boolean().optional().default(false),
  against_blanket_order: z.boolean().optional().default(false),
  blanket_order: z.string().optional(),
  blanket_order_rate: z.number().optional(),
  received_qty: z.number().optional(),
  returned_qty: z.number().optional(),
  billed_amt: z.number().optional(),
  expense_account: z.string().optional(),
  wip_composite_asset: z.string().optional(),
  manufacturer: z.string().optional(),
  manufacturer_part_no: z.string().optional(),
  bom: z.string().optional(),
  include_exploded_items: z.boolean().optional().default(false),
  weight_per_unit: z.number().optional(),
  total_weight: z.number().optional(),
  weight_uom: z.string().optional(),
  project: z.string().optional(),
  cost_center: z.string().optional(),
  is_fixed_asset: z.boolean().optional().default(false),
  item_tax_rate: z.string().optional(),
  production_plan: z.string().optional(),
  production_plan_item: z.string().optional(),
  production_plan_sub_assembly_item: z.string().optional(),
  page_break: z.boolean().optional().default(false),
  job_card: z.string().optional(),
});

export type PurchaseOrderItem = z.infer<typeof PurchaseOrderItemSchema>;

export const PurchaseOrderItemInsertSchema = PurchaseOrderItemSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type PurchaseOrderItemInsert = z.infer<typeof PurchaseOrderItemInsertSchema>;